feat(example): split SectionList and ScrollView host demos
Add tabbed example entry with separate demo files so consumers can
copy either host pattern; share data/theme/ui helpers and update docs.
